fix: Range end Date
Fix the setRange(), because if you use the range e.g. "lastX", the end date must be blank. The Parameter looks like this: "date=lastX" and not "date=lastX,xx-xx-xx". The predefined dates e.g. "lastX" refers to the period e.g. "day". But if you use the period "range" the end Date bust be given.