Improved error response handling
Allows the "_finishResponse" method to to handle parsing the response, which in turn handles validating the response using exceptions to catch issues properly.

Response validation checks the response status code in case it's anything other than 200 so that we can throw exceptions that include the server response information for better troubleshooting.